+++ database/seeders/QuoteSeeder.php
@@ -18,6 +18,8 @@ class QuoteSeeder extends Seeder
             $customer = \App\Models\Partner::all()->random();
             $customer_address = $customer->addresses->first();
             $discount = fake()->randomFloat(2, 0, 100);
+            $payment_method = \App\Models\Category::where('module', 'payment_method')->get()->random();
+
             \App\Models\Quote::create([
                 'id' => $id,
                 'number' => \App\Models\Quote::getQuoteNumber(),
@@ -55,7 +57,7 @@ class QuoteSeeder extends Seeder
                 'notes' => fake()->text(200),
                 'footer' => fake()->text(200),
                 'discount' => $discount,
-                'payment_method' => fake()->randomElement(['cash', 'bank_transfer', 'credit_card', 'paypal', 'stripe', 'other']),
+                'payment_method_id' => $payment_method->id,
                 'discount_type' => 'percent',
                 'total' => 0,
             ]);